+++ ionoscloud/resource_volume.go
@@ -78,36 +78,31 @@ func resourceVolume() *schema.Resource {
 			},
 			"cpu_hot_plug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"ram_hot_plug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"nic_hot_plug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"nic_hot_unplug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"disc_virtio_hot_plug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"disc_virtio_hot_unplug": {
 				Type:     schema.TypeBool,
-				Optional: true,
 				Computed: true,
 			},
 			"backup_unit_id": {
 				Type:     schema.TypeString,
+				Optional: true,
 				Computed: true,
 			},
 			"user_data": {
@@ -260,36 +255,6 @@ func resourceVolumeCreate(ctx context.Context, d *schema.ResourceData, meta inte
 		volume.Properties.LicenceType = nil
 	}
 
-	if v, ok := d.GetOk("cpu_hot_plug"); ok {
-		vBool := v.(bool)
-		volume.Properties.CpuHotPlug = &vBool
-	}
-
-	if v, ok := d.GetOk("ram_hot_plug"); ok {
-		vBool := v.(bool)
-		volume.Properties.RamHotPlug = &vBool
-	}
-
-	if v, ok := d.GetOk("nic_hot_unplug"); ok {
-		vBool := v.(bool)
-		volume.Properties.NicHotPlug = &vBool
-	}
-
-	if v, ok := d.GetOk("nic_hot_unplug"); ok {
-		vBool := v.(bool)
-		volume.Properties.NicHotUnplug = &vBool
-	}
-
-	if v, ok := d.GetOk("disc_virtio_hot_plug"); ok {
-		vBool := v.(bool)
-		volume.Properties.DiscVirtioHotPlug = &vBool
-	}
-
-	if v, ok := d.GetOk("disc_virtio_hot_unplug"); ok {
-		vBool := v.(bool)
-		volume.Properties.DiscVirtioHotUnplug = &vBool
-	}
-
 	if image != "" {
 		volume.Properties.Image = &image
 	} else {
@@ -308,16 +273,29 @@ func resourceVolumeCreate(ctx context.Context, d *schema.ResourceData, meta inte
 		volume.Properties.AvailabilityZone = &raw
 	}
 
-	userData := d.Get("user_data").(string)
-	if userData != "" {
+	if userData, ok := d.GetOk("user_data"); ok {
 		if image == "" && imageAlias == "" {
-			diags := diag.FromErr(fmt.Errorf("it is mandatory to provied either public image that has cloud-init compatibility in conjunction with backup unit id property "))
+			diags := diag.FromErr(fmt.Errorf("it is mandatory to provide either public image that has cloud-init compatibility in conjunction with user_data property "))
 			return diags
 		} else {
+			userData := userData.(string)
 			volume.Properties.UserData = &userData
 		}
-	} else {
-		volume.Properties.UserData = nil
+	}
+
+	if backupUnitId, ok := d.GetOk("backup_unit_id"); ok {
+		if IsValidUUID(backupUnitId.(string)) {
+			if image == "" && imageAlias == "" {
+				diags := diag.FromErr(fmt.Errorf("it is mandatory to provide either public image that has cloud-init compatibility in conjunction with backup_unit_id property "))
+				return diags
+			} else {
+				backupUnitId := backupUnitId.(string)
+				volume.Properties.BackupunitId = &backupUnitId
+			}
+		} else {
+			diags := diag.FromErr(fmt.Errorf("the backup_unit_id that you specified is not a valid UUID"))
+			return diags
+		}
 	}
 
 	volume, apiResponse, err := client.VolumesApi.DatacentersVolumesPost(ctx, dcId).Volume(volume).Execute()
@@ -538,42 +516,14 @@ func resourceVolumeUpdate(ctx context.Context, d *schema.ResourceData, meta inte
 		newValueStr := newValue.(string)
 		properties.AvailabilityZone = &newValueStr
 	}
-	if d.HasChange("cpu_hot_plug") {
-		_, newValue := d.GetChange("cpu_hot_plug")
-		newValueBool := newValue.(bool)
-		properties.CpuHotPlug = &newValueBool
-	}
-	if d.HasChange("ram_hot_plug") {
-		_, newValue := d.GetChange("ram_hot_plug")
-		newValueBool := newValue.(bool)
-		properties.RamHotPlug = &newValueBool
-	}
-	if d.HasChange("nic_hot_plug") {
-		_, newValue := d.GetChange("nic_hot_plug")
-		newValueBool := newValue.(bool)
-		properties.NicHotPlug = &newValueBool
-	}
-
-	if d.HasChange("nic_hot_unplug") {
-		_, newValue := d.GetChange("nic_hot_unplug")
-		newValueBool := newValue.(bool)
-		properties.NicHotUnplug = &newValueBool
-	}
-
-	if d.HasChange("disc_virtio_hot_plug") {
-		_, newValue := d.GetChange("disc_virtio_hot_plug")
-		newValueBool := newValue.(bool)
-		properties.DiscVirtioHotPlug = &newValueBool
-	}
-
-	if d.HasChange("disc_virtio_hot_unplug") {
-		_, newValue := d.GetChange("disc_virtio_hot_unplug")
-		newValueBool := newValue.(bool)
-		properties.DiscVirtioHotUnplug = &newValueBool
-	}
 
 	if d.HasChange("user_data") {
-		diags := diag.FromErr(fmt.Errorf("User data property of resource volume is immutable "))
+		diags := diag.FromErr(fmt.Errorf("user_data property of resource volume is immutable "))
+		return diags
+	}
+
+	if d.HasChange("backup_unit_id") {
+		diags := diag.FromErr(fmt.Errorf("backup_unit_id property of resource volume is immutable "))
 		return diags
 	}